+++ src/ui/hocs/Form.tsx
@@ -1,28 +1,32 @@
 import React from 'react'
-import { useForm, useWatch, FormProvider, useFormContext, useFieldArray } from 'react-hook-form'
+import {
+  useForm,
+  useWatch,
+  FormProvider,
+  useFormContext,
+  useFieldArray,
+} from 'react-hook-form'
 import PropTypes, { InferProps } from 'prop-types'
 
 const propTypes = {
   children: PropTypes.node.isRequired,
   opts: PropTypes.object,
-  onSubmit: PropTypes.func.isRequired
+  onSubmit: PropTypes.func.isRequired,
 }
 
 type Props = InferProps<typeof propTypes>
 
 const defaultProps: Props = {
   opts: {},
-  onSubmit () {}
+  onSubmit() {},
 }
 
-function Form ({ children, opts, onSubmit }: Props) {
+function Form({ children, opts, onSubmit }: Props) {
   const methods = useForm<any>(opts)
 
   return (
     <FormProvider {...methods}>
-      <form onSubmit={methods.handleSubmit(onSubmit)}>
-        {children}
-      </form>
+      <form onSubmit={methods.handleSubmit(onSubmit)}>{children}</form>
     </FormProvider>
   )
 }