+++ src/utils/capitalize.ts
@@ -1,1 +1,3 @@
-export default (string: string = ''): string => string.charAt(0).toUpperCase() + string.slice(1)
+export default function capitalize(string: string = ''): string {
+  return string.charAt(0).toUpperCase() + string.slice(1)
+}